LayoutTests/fast/canvas/webgl/copyBufferSubData-expected.txt

CONSOLE MESSAGE: line 23: TypeError: null is not an object (evaluating 'context.ARRAY_BUFFER')
Make sure that getBufferSubData() doesn't work on WebGL 1 contexts

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL successfullyParsed should be true (of type boolean). Was undefined (of type undefined).

TEST COMPLETE
